+++ FE-Project/src/components/editpost/Edit.tsx
@@ -2,18 +2,18 @@ import axios from "axios";
 import React, { useEffect, useState } from "react";
 import { FindNewPostError } from "../../hooks/Error";
 import NewPostModal from "../modal/NewPostModal";
-import Contact from "./postform/Contact";
-import Deadline from "./postform/Deadline";
-import Duration from "./postform/Duration";
-import Introduce from "./postform/Introduce";
-import Progress from "./postform/Progress";
-import Purpose from "./postform/Purpose";
-import Recruitment from "./postform/Recruitment";
-import RecruitNumber from "./postform/RecruitNumber";
-import TechStack from "./postform/TechStack";
-import TimeAndPlace from "./postform/TimeAndPlace";
-import Title from "./postform/Title";
-import Way from "./postform/Way";
+import Contact from "../newpost/postform/Contact";
+import Deadline from "../newpost/postform/Deadline";
+import Duration from "../newpost/postform/Duration";
+import Introduce from "../newpost/postform/Introduce";
+import Progress from "../newpost/postform/Progress";
+import Purpose from "../newpost/postform/Purpose";
+import Recruitment from "../newpost/postform/Recruitment";
+import RecruitNumber from "../newpost/postform/RecruitNumber";
+import TechStack from "../newpost/postform/TechStack";
+import TimeAndPlace from "../newpost/postform/TimeAndPlace";
+import Title from "../newpost/postform/Title";
+import Way from "../newpost/postform/Way";
 import { Detail } from "../project/ProjectDetail.interface"
 import { getDetailData } from "../../hooks/axios/Project";
 import { useParams } from "react-router-dom";
