+++ FE-Project/src/page/EditPost.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import Header from '../components/header/Header'
-import Edit from '../components/newpost/Edit'
+import Edit from '../components/editpost/Edit'
 import Footer from '../components/footer/Footer'
 
 export default function EditPost() {
